+++ DentalOfficeManagement/src/main/java/com/utn/DentalOfficeManagement/Model/Odontologo.java
@@ -13,7 +13,7 @@ public class Odontologo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_odontologo")
-    private Integer idOdontologo;
+    private Long idOdontologo;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "usuario_id", nullable = false, unique = true)
