+++ DentalOfficeManagement/src/main/java/com/utn/DentalOfficeManagement/Model/Paciente.java
@@ -16,7 +16,7 @@ public class Paciente {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_paciente")
-    private Integer idPaciente;
+    private Long idPaciente;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "id_plan", nullable = false)
