+++ DentalOfficeManagement/src/main/java/com/utn/DentalOfficeManagement/Model/PiezaDental.java
@@ -17,7 +17,7 @@ public class PiezaDental {
 
     @Id
     @Column(name = "id_pieza_dental", unique = true)
-    private Integer idPiezaDental; // No auto-generado: número de pieza es fijo (1-32)
+    private Long idPiezaDental; // No auto-generado: número de pieza es fijo (1-32)
 
     @Column(name = "numero_pieza", nullable = false, precision = 5, scale = 2)
     private BigDecimal numeroPieza;
