+++ DentalOfficeManagement/src/main/java/com/utn/DentalOfficeManagement/Model/Turno.java
@@ -16,7 +16,7 @@ public class Turno {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_turno")
-    private Integer idTurno;
+    private Long idTurno;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "paciente_id", nullable = false)
